code restructure, move firmware update related functions accordingly

+++ common/os_types.h
@@ -44,7 +44,7 @@ struct os_ctxt {
 #endif
     sem_t    fwupd_reply_semid;
     uint8_t  fwupd_reply_cmd;
-    uint16_t fwupd_reply_param;
+    uint32_t fwupd_reply_param;
 
     // For retransmission in case of crc error on the rcp
     // FIXME: rename this and the structure / naive circular buffer : rearch
